refactor: DateFormatter를 .formatted() API로 전환
- Date+Extensions: iOS 15+ .formatted() 방식으로 변경
- String+Extensions: Date에서 String 관련 로직 분리
