+++ StopSun/StopSun/Extensions/Date+Extensions.swift
@@ -11,34 +11,36 @@ import Foundation
 
 extension Date {
     
-    /// API 요청용 날짜 문자열 (yyyy-MM-dd)
+    // MARK: - API용 (고정 포맷)
+     
+     /// API 요청용 날짜 문자열 (yyyy-MM-dd)
+     ///
+     /// ```swift
+     /// Date().toAPIDateString  // "2026-01-30"
+     /// ```
     var toAPIDateString: String {
-        let formatter = DateFormatter()
-        formatter.dateFormat = "yyyy-MM-dd"
-        return formatter.string(from: self)
-    }
-}
-
-// MARK: - String Extensions
-
-extension String {
-    
-    /// WeatherAPI 시간 문자열에서 시간 추출
-    /// "2025-01-27 14:00" → 14
-    var toHour: Int? {
-        let components = self.split(separator: " ")
-        guard components.count >= 2 else { return nil }
-        
-        let timePart = components[1]
-        let hourString = timePart.split(separator: ":").first
-        return hourString.flatMap { Int($0) }
+        formatted(.iso8601.year().month().day().dateSeparator(.dash))
     }
     
-    /// WeatherAPI 시간 문자열을 Date로 변환
-    /// "2025-01-27 14:00" → Date
-    var toDate: Date? {
-        let formatter = DateFormatter()
-        formatter.dateFormat = "yyyy-MM-dd HH:mm"
-        return formatter.date(from: self)
+    // MARK: - UI용 (Localized)
+    
+    /// 시간만 표시 (오후 2:30)
+    var toTimeString: String {
+        formatted(date: .omitted, time: .shortened)
+    }
+    
+    /// 날짜만 표시 (1월 30일)
+    var toDateString: String {
+        formatted(date: .abbreviated, time: .omitted)
+    }
+    
+    /// 날짜 + 시간 표시 (1월 30일 오후 2:30)
+    var toDateTimeString: String {
+        formatted(date: .abbreviated, time: .shortened)
+    }
+    
+    /// 상대 시간 표시 (5분 전, 1시간 전)
+    var toRelativeString: String {
+        formatted(.relative(presentation: .named))
     }
 }
